+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v2.6.0 (2023-04-17) ##
 
+- **Improved:** `json.dump_from_csv` reduce memory consumption, now uses `idump_into`
+- **Added:** `json.idump_lines`
 - **Added:** `json.idump_into`
 - **Added:** `file.get_valid_filename`
 - **Added:** `iterator.partition`